+++ src/components/level/toolbars/EditToolbar/EditToolbar.tsx
@@ -1,14 +1,21 @@
 import { FC } from "react";
 import cn from "classnames";
 import { useStore } from "effector-react";
+import { ReactComponent as MurphyRuns } from "assets/img/murphy-run-right.svg";
+import { TEST_LEVEL_TITLE, TEST_LEVEL_URL } from "configs";
+import { getDriver } from "drivers";
 import {
+  $currentDriverName,
   $currentLevelUndoQueue,
   redoCurrentLevel,
   undoCurrentLevel,
 } from "models/levelsets";
-import { Button, Toolbar, ToolbarSeparator } from "ui/button";
+import { $prefConfirmedTestSO, setPrefAskTestSO } from "models/preferences";
+import { Button, TextButton, Toolbar, ToolbarSeparator } from "ui/button";
+import { ask, msgBox } from "ui/feedback";
 import { svgs } from "ui/icon";
-import { ContainerProps } from "ui/types";
+import { Checkbox } from "ui/input";
+import { ColorType, ContainerProps } from "ui/types";
 import cl from "./EditToolbar.module.scss";
 
 interface Props extends ContainerProps {}
@@ -31,15 +38,99 @@ export const EditToolbar: FC<Props> = ({ className, ...rest }) => {
         onClick={redoCurrentLevel}
       />
       <ToolbarSeparator />
-      <Button icon={<svgs.Cut />} disabled />
-      <Button icon={<svgs.Copy />} disabled />
-      <Button icon={<svgs.Paste />} disabled />
-      <Button icon={<svgs.DeleteSelection />} disabled />
+      {/*<Button icon={<svgs.Cut />} disabled />*/}
+      {/*<Button icon={<svgs.Copy />} disabled />*/}
+      {/*<Button icon={<svgs.Paste />} disabled />*/}
+      {/*<Button icon={<svgs.DeleteSelection />} disabled />*/}
       {/*<Button disabled>PNG</Button>*/}
       {/*<ToolbarSeparator />*/}
       {/*<Button disabled>Copy level to clipboard</Button>*/}
       {/*<Button disabled>Paste level from clipboard</Button>*/}
       {/*<Button disabled>Internal/System clipboard? (via textarea?)</Button>*/}
+      <TextButton
+        onClick={handleTestClick}
+        icon={<MurphyRuns />}
+        title={`Test level at ${TEST_LEVEL_TITLE}`}
+      />
     </Toolbar>
+  );
+};
+
+const handleTestClick = () => {
+  const { writer, createLevelset } = getDriver($currentDriverName.getState()!)!;
+  if (!writer) {
+    return;
+  }
+  const level = $currentLevelUndoQueue.getState()!.current;
+  const [valid, errors] = level.isPlayable();
+  if (!valid) {
+    msgBox(
+      <div>
+        <p>The level is unplayable due to the following:</p>
+        <ul>
+          {errors.map((err, i) => (
+            <li key={i}>{err}</li>
+          ))}
+        </ul>
+      </div>,
+    );
+    return;
+  }
+
+  const url = `${TEST_LEVEL_URL}#${window.btoa(
+    String.fromCharCode(
+      ...new Uint8Array(writer.writeLevelset(createLevelset([level]))),
+    ),
+  )}`;
+
+  if ($prefConfirmedTestSO.getState()) {
+    if (!window.open(url, "_blank")) {
+      msgBox(
+        <div>
+          <p>Could not open new window. Follow the link:</p>
+          <p>
+            <a href={url} target="_blank" rel="noopener noreferrer">
+              Test level at {TEST_LEVEL_TITLE}
+            </a>
+          </p>
+        </div>,
+        { button: { text: "Close" } },
+      );
+    }
+    return;
+  }
+
+  ask(<ConfirmTestSO />, {
+    buttons: {
+      okText: `Go to ${TEST_LEVEL_TITLE} test`,
+      ok: {
+        uiColor: ColorType.SUCCESS,
+        href: url,
+        target: "_blank",
+        rel: "noopener noreferrer",
+      },
+    },
+  });
+};
+
+const ConfirmTestSO: FC = () => {
+  const confirmed = useStore($prefConfirmedTestSO);
+
+  return (
+    <div>
+      <p>
+        To test a level, it <strong>will be sent to</strong>{" "}
+        <a href={TEST_LEVEL_URL} target="_blank" rel="noopener noreferrer">
+          {TEST_LEVEL_TITLE} test page
+        </a>
+        . It will be opened in new tab/window, so you will not loss undo history
+        in current editing session.
+      </p>
+      <div>
+        <Checkbox checked={confirmed} onChange={setPrefAskTestSO}>
+          Don't show this confirmation again
+        </Checkbox>
+      </div>
+    </div>
   );
 };